item1

+++ 02_projet/Item1.java
@@ -3,7 +3,7 @@ public class Item1 {
         int x, y;
         x = 3;
         y = 2;
-        double a = x / y;
+        double a = (double) x / y;
         double f = a * y;
         System.out.println(f);
     }
